backend/admin_backend/requirements.txt
boto3>=1.28.0
python-jose>=3.3.0
PyJWT>=2.6.0
